clip_service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
Pillow>=10.0.0
python-multipart>=0.0.6
chromadb==0.4.24
python-dotenv>=1.0.0
numpy>=1.24.0,<2.0
mysql-connector-python>=8.0.0
sentence-transformers>=2.2.0
open_clip_torch>=2.20.0
